name updates

+++ Visibrick/Services/Sync/LegoColorSyncService.cs
@@ -1,5 +1,5 @@
 using System.Diagnostics;
-using LEGOOO;
+using VisiBrick;
 using System.Globalization;
 using System.IO.Compression;
 using Microsoft.Data.Sqlite;
@@ -28,9 +28,9 @@ public class LegoColorSyncService
 
     
 
-    public async Task SyncColorsAsync(string dbPath)
+    public async Task SyncLegoColorsAsync()
     {
-        var url = "https://cdn.rebrickable.com/media/downloads/colors.csv.gz";
+        var url = "https://cdn.rebrickable.com/media/downloads/colors.csv.gz?1782285120.1112702";
 
         var tempDir = Path.Combine(FileSystem.CacheDirectory, "rebrickable_colors");
         Directory.CreateDirectory(tempDir);
@@ -40,9 +40,6 @@ public class LegoColorSyncService
 
         try
         {
-            // -------------------------
-            // 1. DOWNLOAD FILE
-            // -------------------------
             using (var client = new HttpClient())
             using (var response = await client.GetAsync(url))
             {
@@ -50,11 +47,9 @@ public class LegoColorSyncService
 
                 await using var fs = new FileStream(gzPath, FileMode.Create, FileAccess.Write);
                 await response.Content.CopyToAsync(fs);
+                Debug.WriteLine("Downloaded Colors");
             }
 
-            // -------------------------
-            // 2. DECOMPRESS GZ -> CSV
-            // -------------------------
             await using (var gzStream = File.OpenRead(gzPath))
             using (var decompressionStream = new GZipStream(gzStream, CompressionMode.Decompress))
             await using (var outFile = File.Create(csvPath))
@@ -62,81 +57,37 @@ public class LegoColorSyncService
                 await decompressionStream.CopyToAsync(outFile);
             }
 
-            // -------------------------
-            // 3. OPEN SQLITE
-            // -------------------------
-            using var conn = new SqliteConnection($"Data Source={dbPath}");
-            conn.Open();
+            var colors = new List<LegoColor>();
 
-            using var tx = conn.BeginTransaction();
-
-            // -------------------------
-            // 4. DELETE OLD DATA
-            // -------------------------
-            var deleteCmd = conn.CreateCommand();
-            deleteCmd.CommandText = "DELETE FROM LegoColor";
-            deleteCmd.ExecuteNonQuery();
-
-            // -------------------------
-            // 5. PREPARE INSERT
-            // -------------------------
-            var insertCmd = conn.CreateCommand();
-            insertCmd.CommandText =
-                @"INSERT INTO LegoColor (Id, Name, Rgb, IsTrans)
-                VALUES ($id, $name, $rgb, $isTrans)";
-
-            var pId = insertCmd.CreateParameter(); pId.ParameterName = "$id";
-            var pName = insertCmd.CreateParameter(); pName.ParameterName = "$name";
-            var pRgb = insertCmd.CreateParameter(); pRgb.ParameterName = "$rgb";
-            var pTrans = insertCmd.CreateParameter(); pTrans.ParameterName = "$isTrans";
-
-            insertCmd.Parameters.AddRange(new[]
-            {
-                pId, pName, pRgb, pTrans
-            });
-
-            // -------------------------
-            // 6. READ CSV + INSERT
-            // -------------------------
             var lines = await File.ReadAllLinesAsync(csvPath);
 
-            for (int i = 1; i < lines.Length; i++) // skip header
+            for (int i = 1; i < lines.Length; i++)
             {
                 var cols = lines[i].Split(',');
 
                 if (cols.Length < 4)
                     continue;
 
-                pId.Value = int.Parse(cols[0], CultureInfo.InvariantCulture);
-                pName.Value = cols[1];
-                pRgb.Value = cols[2];
-                pTrans.Value = cols[3] == "t" ? 1 : 0;
-
-                insertCmd.ExecuteNonQuery();
+                colors.Add(new LegoColor
+                {
+                    Id = int.Parse(cols[0], CultureInfo.InvariantCulture),
+                    ColorName = cols[1],
+                    ColorRGB = cols[2],
+                    Is_translucent = cols[3] == "t" ? 1 : 0
+                });
             }
 
-            tx.Commit();
+            Debug.WriteLine("FULL COLORS LIST: " + colors.ToString());
+
+            await _repo.ReplaceAllAsync(colors);
+            
         }
         finally
         {
-            // -------------------------
-            // 7. CLEANUP TEMP FILES
-            // -------------------------
-            try
-            {
-                if (File.Exists(gzPath))
-                    File.Delete(gzPath);
-
-                if (File.Exists(csvPath))
-                    File.Delete(csvPath);
-
-                if (Directory.Exists(tempDir))
-                    Directory.Delete(tempDir, true);
-            }
-            catch
-            {
-                // ignore cleanup failures
-            }
+            if (File.Exists(gzPath)) File.Delete(gzPath);
+            if (File.Exists(csvPath)) File.Delete(csvPath);
+            if (Directory.Exists(tempDir)) Directory.Delete(tempDir, true);
+            Debug.WriteLine("DELETED FILES");
         }
     }
 }
